chore(): cleanups and minor fixes
