+++ apps/akasha/package.json
@@ -13,8 +13,7 @@
   "license": "AGPL-3.0",
   "main": "lib/index.js",
   "directories": {
-    "lib": "lib",
-    "test": "__tests__"
+    "lib": "lib"
   },
   "files": [
     "lib"
@@ -31,15 +30,7 @@
     "prepare": "npm run tsc",
     "pack": "NODE_OPTIONS='--max-old-space-size=4096' webpack --config webpack.config.js",
     "postpack": "copyfiles -f ./dist/* ../../ui/build/apps/akasha",
-    "watch": "nodemon"
-  },
-  "nodemonConfig": {
-    "ignore": [
-      "lib/*",
-      "dist/*"
-    ],
-    "exec": "npm run pack || exit 1",
-    "ext": "ts,tsx"
+    "watch": "webpack --config webpack.config.js --watch"
   },
   "bugs": {
     "url": "https://github.com/AKASHAorg/akasha-world-framework/issues"
@@ -54,7 +45,6 @@
     "@akashaproject/ui-awf-hooks": "*",
     "@akashaproject/ui-widget-feed": "*",
     "@akashaproject/ui-widget-login": "*",
-    "jest": "^26.6.3",
     "react-i18next": "11.8.13",
     "react-router-dom": "5.2.0",
     "single-spa-react": "4.1.1"
@@ -65,8 +55,7 @@
     "@types/react-dom": "17.0.3",
     "@types/react-router-dom": "5.1.7",
     "@types/react-test-renderer": "17.0.1",
-    "@types/single-spa-react": "3.0.1",
-    "nodemon": "2.0.7"
+    "@types/single-spa-react": "3.0.1"
   },
   "peerDependencies": {
     "react": "17.0.2",
