+++ apps/akasha/src/components/__tests__/entry-card-renderer.spec.tsx
@@ -2,29 +2,34 @@ import * as React from 'react';
 import EntryCardRenderer from '../feed-page/entry-card-renderer';
 import { act, create, ReactTestRenderer } from 'react-test-renderer';
 import { cleanup } from '@testing-library/react';
+import DS from '@akashaproject/design-system';
+
+const { ThemeSelector, lightTheme } = DS;
 
 describe('<ContentCard /> component', () => {
   let componentWrapper: ReactTestRenderer = create(<></>);
   beforeEach(() => {
     act(() => {
       componentWrapper = create(
-        <EntryCardRenderer
-          sdkModules={{}}
-          logger={{ log: console.log }}
-          globalChannel={{}}
-          ethAddress=""
-          onBookmark={jest.fn}
-          onAvatarClick={jest.fn}
-          onMentionClick={jest.fn}
-          onNavigate={jest.fn}
-          onRepost={jest.fn}
-          onTagClick={jest.fn}
-          onFlag={jest.fn}
-          onLinkCopy={jest.fn}
-          sharePostUrl=""
-          rxjsOperators={{}}
-          singleSpaNavigate={jest.fn}
-        />,
+        <ThemeSelector availableThemes={[lightTheme]} settings={{ activeTheme: 'lightTheme' }}>
+          <EntryCardRenderer
+            sdkModules={{}}
+            logger={{ log: console.log }}
+            globalChannel={{}}
+            ethAddress=""
+            onBookmark={jest.fn}
+            onAvatarClick={jest.fn}
+            onMentionClick={jest.fn}
+            onNavigate={jest.fn}
+            onRepost={jest.fn}
+            onTagClick={jest.fn}
+            onFlag={jest.fn}
+            onLinkCopy={jest.fn}
+            sharePostUrl=""
+            rxjsOperators={{}}
+            singleSpaNavigate={jest.fn}
+          />
+        </ThemeSelector>,
       );
     });
   });
