+++ ui/design/src/styles/themes/utils/create-theme.ts
@@ -1,4 +1,4 @@
-import { DefaultTheme } from 'styled-components';
+import { DefaultTheme } from '../interfaces';
 import breakpoints from '../../responsive-breakpoints';
 import shapes from '../../shapes';
 import colors from '../light/colors-light';
@@ -6,11 +6,13 @@ import createGrommetTheme from './create-grommet-theme';
 
 // defaults to 'light theme' if no override is passed
 export function createTheme(overrides?: object): DefaultTheme {
+
   const styles = {
     colors,
     shapes,
     breakpoints: breakpoints.global.breakpoints,
     ...overrides,
   };
+
   return createGrommetTheme(styles);
 }
